[GAME] Added game particles.
Added particle objects and a collection of particles. 
Added some functions to convert from cartesian to screen coords. 
Added toy functions for particles in the main class. 
Modified Point class to contain doubles.

+++ src/graphics/Point.h
@@ -1,7 +1,7 @@
 #pragma once
 namespace Gravity {
 	struct Point {
-		int x, y;
-		Point(int inputX, int inputY) : x(inputX), y(inputY) {};
+		double x, y;
+		Point(double inputX, double inputY) : x(inputX), y(inputY) {};
 	};
 }
